More MantidPlot qwt6 porting.
Refs #0000

+++ Code/Mantid/MantidPlot/src/SmoothCurveDialog.cpp
@@ -148,6 +148,6 @@ void SmoothCurveDialog::activateCurve(const QString& curveName)
 	if (!c || c->rtti() != QwtPlotItem::Rtti_PlotCurve)
 		return;
 
-	boxPointsLeft->setMaxValue(c->dataSize()/2);
+	boxPointsLeft->setMaxValue(static_cast<int>(c->dataSize())/2);
 	}
 }
